# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: XHS AI Tool API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sensitive-words:
    get:
      operationId: listSensitiveWords
      tags: [ai]
      summary: List sensitive word rules
      responses:
        "200":
          description: Sensitive words list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SensitiveWord"
    post:
      operationId: createSensitiveWord
      tags: [ai]
      summary: Add a sensitive word rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSensitiveWordBody"
      responses:
        "201":
          description: Sensitive word added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SensitiveWord"
  /sensitive-words/{id}:
    delete:
      operationId: deleteSensitiveWord
      tags: [ai]
      summary: Delete a sensitive word rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    SensitiveWord:
      type: object
      properties:
        id:
          type: integer
        word:
          type: string
        category:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - word
        - category
        - severity
        - createdAt
    CreateSensitiveWordBody:
      type: object
      properties:
        word:
          type: string
        category:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
      required:
        - word
        - category
        - severity
...