# lib/api-zod/src/generated/types/createSensitiveWordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * XHS AI Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSensitiveWordBodySeverity } from "./createSensitiveWordBodySeverity";

export interface CreateSensitiveWordBody {
  word: string;
  category: string;
  severity: CreateSensitiveWordBodySeverity;
}